refactor(backend): move inputs to module
